+++ frontend-hybrid/yakal/lib/main.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
+import 'package:yakal/widgets/Base/my_bottom_navigation_bar.dart';
 
 void main() async {
-  // Load environment variables
-  // await dotenv.load(fileName: 'assets/config/.env');
-
   // Setup splash
   WidgetsBinding widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
   FlutterNativeSplash.preserve(widgetsBinding: widgetsBinding);
@@ -18,88 +15,17 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    // Remove splash
     FlutterNativeSplash.remove();
+
+    // 앱 실행
     return MaterialApp(
       title: '약 알',
       theme: ThemeData(
         useMaterial3: true,
       ),
+      // 하단 탭
       home: const MyBottomNavigationBar(),
-    );
-  }
-}
-
-class MyBottomNavigationBar extends StatefulWidget {
-  const MyBottomNavigationBar({Key? key}) : super(key: key);
-
-  @override
-  _MyBottomNavigationBarState createState() => _MyBottomNavigationBarState();
-}
-
-class _MyBottomNavigationBarState extends State<MyBottomNavigationBar> {
-  int _currentIndex = 0;
-
-  final List<Widget> _screens = <Widget>[
-    const HomeTab(),
-    const MyPageTab(),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('약 알'),
-      ),
-      body: _screens[_currentIndex],
-      bottomNavigationBar: Theme(
-        data: ThemeData(
-          canvasColor: Colors.white, // Set the background color to white
-        ),
-        child: BottomNavigationBar(
-          currentIndex: _currentIndex,
-          items: const <BottomNavigationBarItem>[
-            BottomNavigationBarItem(
-              icon: Icon(Icons.home),
-              label: '홈',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.person),
-              label: '마이페이지',
-            ),
-          ],
-          onTap: (index) {
-            setState(() {
-              _currentIndex = index;
-            });
-          },
-        ),
-      ),
-    );
-  }
-}
-
-class HomeTab extends StatelessWidget {
-  const HomeTab({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Text(
-        'This is the Home Tab',
-      ),
-    );
-  }
-}
-
-class MyPageTab extends StatelessWidget {
-  const MyPageTab({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Text(
-        'This is the MyPage Tab',
-      ),
     );
   }
 }
