+++ frontend-hybrid/yakal/lib/screens/Home/home_screen.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('홈'),
+    );
+  }
+}
