+++ frontend-hybrid/yakal/lib/screens/Profile/profile_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class ProfileScreen extends StatelessWidget {
+  const ProfileScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        'profile',
+      ),
+    );
+  }
+}
